Dividing functions up
smaller consumable bits in prep to add deeper build to include alignment, OCC, etc.
