+++ docker/db-dump/initDb.sql
@@ -80,16 +80,16 @@ values ('Human'), ('Elf'), ('Dwarf'), ('Gnome'), ('Troglodyte'), ('Kobold'), ('G
 Insert into RaceAttributes (RaceId, IQ, IQBonus, ME, MEBonus, MA, MABonus, PS, PSBonus, PP, PPBonus, PE, PEBonus,
 	PB, PBBonus, Spd, SpdBonus, PPE, PPEBonus, Alignment, SpdDig, SpdDigBonus)
 values 	(1,3,0,3,0,3,0,3,0,3,0,3,0,3,0,3,0,2,0,'Any, usually lean toward good and selfish',0,0),
-		(2,3,1,3,0,2,0,3,0,4,0,3,0,5,0,3,0,2,0,'Any, usually lean toward good and selfish',0,0),
-		(3,3,0,3,0,2,0,4,6,3,0,4,0,2,2,2,0,2,0,'Any, usually lean toward good and selfish',1,0),
-		(4,3,0,1,6,3,4,1,4,4,0,3,6,4,0,2,0,2,0,'Any, but most tend to be good or selfish; an evil gnome is a rarity',1,0),
-		(5,2,0,2,0,3,0,4,4,3,6,3,0,2,0,6,0,2,0,'Typically anarchist or evil; but most player characters are likely to be unprincipled, anarchist, aberrant or even good',1,0),
-		(6,3,0,2,0,3,0,3,3,3,0,3,0,1,6,3,0,4,0,'Typically anarchist or evil, but most player characters are likely to be unprincipled, anarchist, aberrant or even good',1,0),
-    (7,2,0,3,0,2,0,3,0,3,6,3,0,2,0,3,0,6,0,'Typically anarchist or evil; but most player characters are likely to be unprincipled, anarchist, aberrant or even good',1,0),
-    (8,2,0,3,6,2,0,3,0,3,0,3,0,2,0,3,0,4,0,'Typically anarchist or evil, but most player characters are likely to be unprincipled, anarchist, aberrant or even good',1,0),
-    (9,2,0,2,0,3,0,3,8,3,0,3,2,2,0,3,0,2,0,'Typically anarchist or evil, but most player characters are likely to be unprincipled, anarchist, aberrant or even good',0,0),
-	  (10,3,0,3,0,2,0,4,4,3,0,3,6,2,0,3,0,3,0,'Typically anarchist or evil, but most player characters are likely to be unprincipled, anarchist, aberrant or even good',0,0),
-	  (11,3,0,2,0,2,0,4,10,4,0,3,6,1,4,2,0,3,0,'Typically anarchist or evil, but most player characters are likely to be unprincipled, anarchist, aberrant or even good',0,0),
-	  (12,3,0,4,6,4,0,3,0,3,0,2,0,2,0,2,0,5,0,'Any',0,0),
-	  (13,3,0,3,0,2,0,4,1,3,0,3,0,3,0,4,0,3,0,'Any, but tend toward principled and aberrant, both alignments with a strong personal code of honor',0,0),
-	  (14,3,0,3,0,2,0,3,1,4,1,3,0,3,0,3,0,3,0,'Any, but tend toward anarchist and miscreant; the antithesis of the noble Wolfen',0,0);
+        (2,3,1,3,0,2,0,3,0,4,0,3,0,5,0,3,0,2,0,'Any, usually lean toward good and selfish',0,0),
+        (3,3,0,3,0,2,0,4,6,3,0,4,0,2,2,2,0,2,0,'Any, usually lean toward good and selfish',1,0),
+        (4,3,0,1,6,3,4,1,4,4,0,3,6,4,0,2,0,2,0,'Any, but most tend to be good or selfish; an evil gnome is a rarity',1,0),
+        (5,2,0,2,0,3,0,4,4,3,6,3,0,2,0,6,0,2,0,'Typically anarchist or evil; but most player characters are likely to be unprincipled, anarchist, aberrant or even good',1,0),
+        (6,3,0,2,0,3,0,3,3,3,0,3,0,1,6,3,0,4,0,'Typically anarchist or evil, but most player characters are likely to be unprincipled, anarchist, aberrant or even good',1,0),
+        (7,2,0,3,0,2,0,3,0,3,6,3,0,2,0,3,0,6,0,'Typically anarchist or evil; but most player characters are likely to be unprincipled, anarchist, aberrant or even good',1,0),
+        (8,2,0,3,6,2,0,3,0,3,0,3,0,2,0,3,0,4,0,'Typically anarchist or evil, but most player characters are likely to be unprincipled, anarchist, aberrant or even good',1,0),
+        (9,2,0,2,0,3,0,3,8,3,0,3,2,2,0,3,0,2,0,'Typically anarchist or evil, but most player characters are likely to be unprincipled, anarchist, aberrant or even good',0,0),
+        (10,3,0,3,0,2,0,4,4,3,0,3,6,2,0,3,0,3,0,'Typically anarchist or evil, but most player characters are likely to be unprincipled, anarchist, aberrant or even good',0,0),
+        (11,3,0,2,0,2,0,4,10,4,0,3,6,1,4,2,0,3,0,'Typically anarchist or evil, but most player characters are likely to be unprincipled, anarchist, aberrant or even good',0,0),
+        (12,3,0,4,6,4,0,3,0,3,0,2,0,2,0,2,0,5,0,'Any',0,0),
+        (13,3,0,3,0,2,0,4,1,3,0,3,0,3,0,4,0,3,0,'Any, but tend toward principled and aberrant, both alignments with a strong personal code of honor',0,0),
+        (14,3,0,3,0,2,0,3,1,4,1,3,0,3,0,3,0,3,0,'Any, but tend toward anarchist and miscreant; the antithesis of the noble Wolfen',0,0);
