+++ main.go
@@ -1,7 +1,22 @@
 package main
 
-import ()
+import (
+	"fmt"
+	"strings"
+)
 
 func main() {
-	builder()
+	keepGoing  := "Y"
+
+	fmt.Println("Welcome to Palladium Fantasy Character Builder!\n")
+	for strings.ToLower(keepGoing)=="y" {
+		startChoice := starter()
+		if startChoice=="2" {
+			builder()
+		} else {
+			reviewer(startChoice)
+		}
+		fmt.Printf("\nWould you like to continue reviewing or building characters? (Y/n) : ")
+		fmt.Scanln(&keepGoing)
+	}
 }
